+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Unreleased
 
-[...]
+[..]
+
+# v7.0.0 (20/06/2019)
+
+- **[BREAKING CHANGE]** Removed `Modal` large prop and added `size` with 3 preset values instead
+
+# v6.0.2 (18/06/2019)
+
+- **[FIX]** Fix `AutoComplete` key stringify and allow customization from outside
 
 - **[FIX]** Remove extra attributes on `Item`-like components when `href` attribute is a string